+++ ui/pages/organismes/[organismeId]/effectifs.tsx
@@ -5,27 +5,28 @@ import Head from "next/head";
 import Page from "@/components/Page/Page";
 import withAuth from "@/components/withAuth";
 import { getAuthServerSideProps } from "@/common/SSR/getAuthServerSideProps";
-import ConfigurationAPI from "@/modules/mon-espace/effectifs/ConfigurationAPI";
+import EffectifsPage from "@/modules/mon-espace/effectifs/EffectifsPage";
 import { useEffectifsOrganisme } from "@/modules/mon-espace/effectifs/useEffectifsOrganisme";
 import { useRouter } from "next/router";
 
 export const getServerSideProps = async (context) => ({ props: { ...(await getAuthServerSideProps(context)) } });
 
-const PageAideConfigurationErpSonOrganisme = () => {
+const PageEffectifsDeSonOrganisme = () => {
   const router = useRouter();
-  const { organisme } = useEffectifsOrganisme(router.query.id as string);
+  const { organisme } = useEffectifsOrganisme(router.query.organismeId as string);
+
   return (
     <Page>
       <Head>
-        <title>Configuration ERP</title>
+        <title>Ses effectifs</title>
       </Head>
       <Box w="100%" pt={[4, 6]} px={[1, 1, 2, 4]} mb={16}>
         <Container maxW="xl" px={0}>
-          {organisme && <ConfigurationAPI organisme={organisme} isMine={false} />}
+          {organisme && <EffectifsPage isMine={false} />}
         </Container>
       </Box>
     </Page>
   );
 };
 
-export default withAuth(PageAideConfigurationErpSonOrganisme);
+export default withAuth(PageEffectifsDeSonOrganisme);
