+++ ui/pages/organismes/[organismeId]/effectifs/televersement.tsx
@@ -5,28 +5,25 @@ import Head from "next/head";
 import Page from "@/components/Page/Page";
 import withAuth from "@/components/withAuth";
 import { getAuthServerSideProps } from "@/common/SSR/getAuthServerSideProps";
-import EffectifsPage from "@/modules/mon-espace/effectifs/EffectifsPage";
 import { useEffectifsOrganisme } from "@/modules/mon-espace/effectifs/useEffectifsOrganisme";
+import Televersements from "@/modules/mon-espace/effectifs/Televersements";
 import { useRouter } from "next/router";
 
 export const getServerSideProps = async (context) => ({ props: { ...(await getAuthServerSideProps(context)) } });
 
-const PageEffectifsDeSonOrganisme = () => {
+const PageImportEffectifsDeSonOrganisme = () => {
   const router = useRouter();
-  const { organisme } = useEffectifsOrganisme(router.query.id as string);
-
+  const { organisme } = useEffectifsOrganisme(router.query.organismeId as string);
   return (
     <Page>
       <Head>
-        <title>Ses effectifs</title>
+        <title>Ses effectifs - Import</title>
       </Head>
       <Box w="100%" pt={[4, 6]} px={[1, 1, 2, 4]} mb={16}>
-        <Container maxW="xl" px={0}>
-          {organisme && <EffectifsPage isMine={false} />}
-        </Container>
+        <Container maxW="xl">{organisme && <Televersements organisme={organisme} />}</Container>
       </Box>
     </Page>
   );
 };
 
-export default withAuth(PageEffectifsDeSonOrganisme);
+export default withAuth(PageImportEffectifsDeSonOrganisme);
